Extract folder tree state
